Add perf metrics
